authorize

+++ app/controllers/speaker/id_cards_controller.rb
@@ -5,6 +5,7 @@ class Speaker::IdCardsController < ApplicationController
   def create
     if params[:id_card]
       id_card = @repayment.id_cards.new(id_card_params)
+      authorize [:speaker, id_card]
       id_card.description = "Id Document"
       if id_card.save
         flash[:notice] = "L'allegato è stato salvato."
@@ -18,6 +19,7 @@ class Speaker::IdCardsController < ApplicationController
 
   def destroy
     id_card = @repayment.id_cards.find(params[:id])
+    authorize [:speaker, id_card]
     id_card.destroy
     redirect_to edit_speaker_repayment_path(@repayment.spkr_token)
   end
